Phase 1: satellites, propagated and cross-validated
Adds orbital mechanics to argus-core and the CelesTrak driver on top of it.

The satellites driver is shaped differently from the others and the difference
matters: fetch and observation run on separate clocks. Element sets are cached
for six hours because CelesTrak asks not to be polled hard for them; positions
are produced every minute by propagating those cached elements. So every
observation here is Quality::Modeled — nothing is measured, and the client must
be able to say so.

The transform chain is the risky part, because getting the sidereal angle wrong
fails quietly: the orbit still looks like an orbit, it is just in the wrong
place, drifting rather than obviously broken. Unit tests can only prove the ISS
lands in a plausible altitude band at a plausible inclination, which a wrong
rotation direction also satisfies. So there is a network-gated cross-check
against an independent tracker, propagating to the reference's own timestamp
since 7.6 km/s makes any looser comparison meaningless. It agrees to 0.1 km.

Two things found while running it:

- CelesTrak 403s clients that pull its full catalogue too often, and the HTTP
  client mapped 403 to Auth, which is terminal — permanently disabling a source
  that sends no credential at all and would have recovered within the hour. 403
  is genuinely ambiguous, so it is now its own error and the scheduler resolves
  it: a source that authenticates has a credential problem, a source that does
  not is being throttled.

- GROUP=active is 16,469 objects. At a 60-second cadence that is ~23 million
  rows a day, for positions perfectly reconstructable from the element sets.
  Narrowed to a curated ~1,100-object core catalogue across every orbital
  regime worth watching; Starlink stays opt-in for the same reason.

Verified live: 948 satellites and 260 earthquakes in the DVR, both sources
reporting live. EntityId::satellite widened to u64 — the catalogue has outgrown
the 5-digit range and truncating would merge two objects onto one track.

85 tests, clippy clean.

+++ crates/argus-core/src/entity.rs
@@ -110,7 +110,11 @@ impl EntityId {
         Self::new(EntityKind::Vessel, mmsi.to_string())
     }
 
-    pub fn satellite(norad_id: u32) -> Self {
+    /// NORAD catalogue number. Takes `u64` rather than `u32` deliberately: the
+    /// public catalogue has outgrown the historical 5-digit range, and
+    /// truncating an id would silently merge two different objects onto one
+    /// track.
+    pub fn satellite(norad_id: u64) -> Self {
         Self::new(EntityKind::Satellite, norad_id.to_string())
     }
 }
